+++ front/lib/api/datasets.ts
@@ -1,5 +1,5 @@
 import { Authenticator } from "@app/lib/auth";
-import { CoreAPI } from "@app/lib/dust_api";
+import { CoreAPI } from "@app/lib/core_api";
 import { Dataset } from "@app/lib/models";
 import { AppType } from "@app/types/app";
 import { DatasetType } from "@app/types/dataset";
