+++ front/pages/api/v1/w/[wId]/apps/[aId]/runs/[runId]/index.ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from "next";
 
 import { getApp } from "@app/lib/api/app";
 import { Authenticator, getAPIKey } from "@app/lib/auth";
-import { CoreAPI } from "@app/lib/dust_api";
+import { CoreAPI } from "@app/lib/core_api";
 import { ReturnedAPIErrorType } from "@app/lib/error";
 import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
