+++ front/pages/api/w/[wId]/apps/[aId]/datasets/index.ts
@@ -3,8 +3,8 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { getApp } from "@app/lib/api/app";
 import { getDatasets } from "@app/lib/api/datasets";
 import { Authenticator, getSession } from "@app/lib/auth";
+import { CoreAPI } from "@app/lib/core_api";
 import { checkDatasetData } from "@app/lib/datasets";
-import { CoreAPI } from "@app/lib/dust_api";
 import { ReturnedAPIErrorType } from "@app/lib/error";
 import { Dataset } from "@app/lib/models";
 import { apiError, withLogging } from "@app/logger/withlogging";
