+++ front/pages/api/w/[wId]/data_sources/[name]/documents/index.ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from "next";
 
 import { getDataSource } from "@app/lib/api/data_sources";
 import { Authenticator, getSession } from "@app/lib/auth";
-import { CoreAPI } from "@app/lib/dust_api";
+import { CoreAPI } from "@app/lib/core_api";
 import { withLogging } from "@app/logger/withlogging";
 import { DocumentType } from "@app/types/document";
