+++ front/types/document.ts
@@ -1,3 +1,3 @@
-import { CoreAPIDocument } from "@app/lib/dust_api";
+import { CoreAPIDocument } from "@app/lib/core_api";
 
 export type DocumentType = CoreAPIDocument;
